+++ changelog.md
@@ -1,3 +1,16 @@
+# version 2.1.0
+
+## features
+
+* Model view (in memory) can be not equal to Insert view (data to insert in DB). For passwords and Date issues.
+* Text field now accepts transformer.
+* Datetime field now writes down Datetime in ISO format as string by default.
+
+
+## Bug fixes
+
+* All tests repaired.
+
 # version 2.0.1
 
 ## features